data: Minor tidying.
